Changed MapBuilder to use ints.

+++ Main/MapBuilder.cs
@@ -36,7 +36,7 @@ namespace Main
                 int height = f.Segments.Count;
 
                 Structure.Floor floor = new Structure.Floor();
-                floor.Setup((uint)width, (uint)height, 3);
+                floor.Setup(width, height, 3);
 
                 for (int row = 0; row < f.Segments.Count; row++)
                 {
@@ -45,7 +45,7 @@ namespace Main
                         var segment = f.Segments[row][col];
                         if (segment.Type != SegmentType.FLOOR) continue;
 
-                        floor.SetFloor((uint)row, (uint)col, (uint)segment.Capacity);
+                        floor.SetFloor(row, col, segment.Capacity);
 
                         var sides = segment.GetSideElements();
 
@@ -57,10 +57,10 @@ namespace Main
                             Structure.WallPlace place = (s == Direction.UP || s == Direction.DOWN) ? Structure.WallPlace.TOP : Structure.WallPlace.LEFT;
 
                             if (sideElement.Type == SideElementType.WALL)
-                                floor.SetWall((uint)r, (uint)c, place);
+                                floor.SetWall(r, c, place);
 
                             if (sideElement.Type == SideElementType.DOOR)
-                                floor.SetDoor((uint)r, (uint)c, (uint)sideElement.Capacity, place);
+                                floor.SetDoor(r, c, sideElement.Capacity, place);
                         }
                     }
                 }
@@ -71,7 +71,7 @@ namespace Main
             // Stairs
             foreach (var stairsPair in _building.Stairs)
             {
-                Structure.Stairs stairs = new Structure.Stairs((uint)stairsPair.First.Capacity, (uint)stairsPair.First.Delay);
+                Structure.Stairs stairs = new Structure.Stairs(stairsPair.First.Capacity, stairsPair.First.Delay);
 
                 // row, col, floor (level)
                 int r, c, f;
@@ -84,8 +84,8 @@ namespace Main
                 c = first.Orientation == Direction.RIGHT ? first.Col + 1 : first.Col;
                 pos = (first.Orientation == Direction.LEFT || first.Orientation == Direction.RIGHT) ? Structure.WallPlace.LEFT : Structure.WallPlace.TOP;
 
-                Structure.WallElementPosition wep1 = new Structure.WallElementPosition((uint)f, (uint)r, (uint)c, pos);
-                Structure.StairsEntry se1 = new Structure.StairsEntry((uint)first.Capacity, wep1);
+                Structure.WallElementPosition wep1 = new Structure.WallElementPosition(f, r, c, pos);
+                Structure.StairsEntry se1 = new Structure.StairsEntry(first.Capacity, wep1);
 
                 // Second stairs entry.
                 var second = stairsPair.Second;
@@ -94,8 +94,8 @@ namespace Main
                 c = second.Orientation == Direction.RIGHT ? second.Col + 1 : second.Col;
                 pos = (second.Orientation == Direction.LEFT || second.Orientation == Direction.RIGHT) ? Structure.WallPlace.LEFT : Structure.WallPlace.TOP;
 
-                Structure.WallElementPosition wep2 = new Structure.WallElementPosition((uint)f, (uint)r, (uint)c, pos);
-                Structure.StairsEntry se2 = new Structure.StairsEntry((uint)second.Capacity, wep2);
+                Structure.WallElementPosition wep2 = new Structure.WallElementPosition(f, r, c, pos);
+                Structure.StairsEntry se2 = new Structure.StairsEntry(second.Capacity, wep2);
 
                 stairs.SetEntries(se1, se2);
 
@@ -119,7 +119,7 @@ namespace Main
                         var segment = f.Segments[row][col];
                         if (segment.PeopleCount > 0)
                         {
-                            Structure.PeopleGroup group = new Structure.PeopleGroup(f.Level, (uint)row, (uint)col, (uint)segment.PeopleCount);
+                            Structure.PeopleGroup group = new Structure.PeopleGroup(f.Level, row, col, segment.PeopleCount);
                             result.People.Add(group);
                         }
